Now you can enter the world with your character; [Game] added UccManager by Gene; [Packets] found some packets; [Game] fixed some packets.

+++ AAEmu.Game/Core/Packets/G2C/SCTimeOfDayPacket.cs
@@ -5,16 +5,17 @@ namespace AAEmu.Game.Core.Packets.G2C
 {
     public class SCTimeOfDayPacket : GamePacket
     {
-        private readonly float _time;
+        private readonly ushort _time;
 
         public SCTimeOfDayPacket(float time) : base(SCOffsets.SCTimeOfDayPacket, 5)
         {
-            _time = time;
+            _time = (ushort)time;
         }
 
         public override PacketStream Write(PacketStream stream)
         {
             stream.Write(_time);
+
             return stream;
         }
     }
